\name{actcal}
\docType{data}
\alias{actcal}
\title{Example data set: Activity calendar from the Swiss Household Panel}
\description{
This data set contains a sample of individual monthly activity statuses from
January to December 2000. The column (variable) names are `jan00', `feb00', etc...
}
\usage{data(actcal)}
\format{A data frame containing 2000 rows and 12 variables.}
\source{Swiss Household Panel}
\references{\url{www.swisspanel.ch}}
\keyword{datasets}
