\name{biofam}
\docType{data}
\alias{biofam}
\title{Example data set: Family life states from the Swiss Household Panel retrospective biographical survey}
\description{
The \emph{biofam} data set was constructed by M\"{u}ller et al. (2007)
from the data of the retrospective biographical survey carried out by the Swiss Household Panel in 2002. In includes only individuals who were at least 30 years old at the time of the survey for whom we consider sequences of family life states between ages 15 and 30. The \emph{biofam} data set describes thus family life courses of 2000 individuals born between 1909 and 1972.}
\usage{data(biofam)}
\format{A data frame containing 2000 rows and 16 variables.}
\source{Swiss Household Panel}
\references{
  M\"{u}ller, N. S., M. Studer, G. Ritschard (2007).
  Classification de parcours de vie \`{a} l'aide de l'optimal matching.
  In \emph{XIVe Rencontre de la Soci\'{e}t\'{e} francophone de classification
  ({SFC} 2007), Paris, 5 - 7 septembre 2007}, pp.\  157--160.
}
\keyword{datasets}
