\name{mvad}
\docType{data}
\alias{mvad}
\title{Example data set: Transition from school to work}
\description{
The data comes from a study by McVicar and Anyadike-Danes on transition from school to work. The data consist of static background characteristics and a time series sequence of 72 monthly labour market activities for each of 712 individuals in a cohort survey. The individuals were followed up from July 1993 to June 1999.}
\usage{data(mvad)}
\format{A data frame containing 712 rows and 86 variables.}
\source{McVicar, Duncan and Anyadike-Danes, Michael}
\references{
McVicar, Duncan and Anyadike-Danes, Michael. Predicting Successful and Unsuccessful Transitions from School to Work by Using Sequence Methods, \emph{Journal of the Royal Statistical Society. Series A (Statistics in
	Society)}, 2002, 165, 2, pp. 317--334.
}
\keyword{datasets}
