\name{read.tda.mdist}
\alias{read.tda.mdist}
\title{Read a distance matrix produced by TDA.}
\description{This function reads a distance matrix produced by TDA into an R object. When computing OM distances in TDA, the output is a 'half' matrix stored in a text file as a vector.}
\usage{
read.tda.mdist(file)
}
\arguments{
  \item{file}{the path to the file containing TDA output.}
}
\value{a R matrix containing the distances.} 
%\seealso{\code{\link{}}
%\examples{}
\keyword{manip}
