\name{TraMineR-internal}
\alias{[.stslist}
\alias{entropy}
\alias{levenshtein}
\alias{LCPdist}
\alias{LCSdist}
\alias{Math.stslist}
\alias{nsubs}
\alias{sconc}
\alias{seqasnum}
\alias{seqshift}
\alias{seqtrunc}
\alias{seqxtract}
\alias{SPELL_to_STS}
\alias{SPS_to_STS}
\alias{STS_to_DSS}
\alias{STS_to_SPS}
\alias{STS_to_SRS}
\alias{STS_to_TSE}
\alias{statefpos}
\alias{turb}
\alias{realvar}
\alias{seqgbar}
\alias{seqcmpinertia}
\title{Internal TraMineR functions}
\description{
 Internal TraMineR functions
}
\details{
Internal functions that are not to be called by the user.
}
\keyword{internal}
