\name{TraMineR.checkupdates}
% \Rdversion{1.1}
\alias{TraMineR.checkupdates}
\title{
Check for TraMineR updates
}

\description{ Check if the installed version of TraMineR is
up-to-date. This function only prints a message and does not need any
argument. It connects to the TraMineR webserver 
(\url{http://mephisto.unige.ch/TraMineR}). }

\value{
Return your current version number of TraMineR and the latest stable and
development version number if more recent versions are available.
}

\usage{
TraMineR.checkupdates()
}
% \examples{
% }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
