\name{TraMineRInternalLayout}
\alias{TraMineRInternalNodeInit}
\alias{TraMineRInternalSplitInit}
\alias{TraMineRInternalLayout}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Access to TraMineR internal functions
}
\description{
Functions allowing other packages to access some TraMineR internal functions. Corresponding functions respectively are \code{DTNInit}, \code{DTNsplit} and \code{TraMineR.setlayout}. For experts only.
}
\usage{
TraMineRInternalNodeInit(...)
TraMineRInternalSplitInit(...)
TraMineRInternalLayout(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{Arguments passed to or from other methods.}
}
