\name{disstreeleaf}
\alias{disstreeleaf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Terminal node membership}
\description{
  Return a factor with the terminal node membership of each case.
}
\usage{
disstreeleaf(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{The tree, a \code{disstree} or \code{DissTreeNode} object.}
}


\seealso{\code{\link{disstree}} for examples}
