\name{disstreeleaf}
\alias{disstreeleaf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Terminal node appartenance}
\description{
  Return a factor with the terminal node appartenance of each cases.
}
\usage{
disstreeleaf(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{The tree}
}


\seealso{\code{\link{disstree}} for examples}
