\name{ex1}
\docType{data}
\alias{ex1}
\title{Example data set with missing values and weights}

\description{
Example data set used to demonstrate the handling of missing values and weights.

The state columns (variable) are named \code{'P1'} to \code{'P13'}.

The alphabet is made of four possible states: \code{A, B, C} and \code{D}.
\cr

The data set contains also case weights (variable \code{weights}). The sum of the weights is 60.
}
\usage{data(ex1)}

\format{A data frame with 7 rows, 13 state variables, 1 weight variable.}

\source{The brain of the TraMineR package maintainer.}

\keyword{Datasets}
