\name{internalLayout}
\alias{internalNodeInit}
\alias{internalSplitInit}
\alias{internalLayout}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Access to internal functions
}
\description{
Functions allowing other packages to access some TraMineR internal functions. For experts only.
}
\usage{
internalNodeInit(...)
internalSplitInit(...)
internalLayout(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{Arguments passed to or from other methods.}
}
