\name{is.seqe}
\alias{is.seqe}
\alias{is.seqelist}
\alias{print.seqe}
\alias{print.seqelist}
\alias{levels.seqelist}
\alias{[.seqelist}
\alias{levels.seqe}
\alias{Math.seqe}
\alias{Math.seqelist}
\alias{Ops.seqe}
\alias{Ops.seqelist}
\alias{Summary.seqe}
\alias{Summary.seqelist}


\title{Event sequence object}
\description{
  TraMineR uses an internal event sequence object for all its operations
  with event sequences such as mining frequent subsequences. The
  function \code{is.seqelist} checks wether the argument is an event
  sequence object while \code{is.seqe} checks wether the argument is a
  single element of an event sequence object.
  %\code{as.seqelist} transform a list of event sequences into an internal event sequence object.

  There is a print method that can be applied to such event sequence object.
}
\usage{
is.seqe(s)
is.seqelist(s)
}
\arguments{
\item{s}{Sequence}
}
\seealso{\code{\link{str.seqelist}}, \code{\link{str.seqe}},\code{\link{as.character.seqelist}} and \code{\link{as.character.seqe}} to convert sequences from an event sequence object into character strings}
%\examples{}

\author{Matthias Studer (with Gilbert Ritschard for the help page)}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Event sequences}
\keyword{internal}
