\name{TraMineRInternal}
\alias{TraMineRInternalLayout}
\alias{TraMineRInternalLegend}
\alias{TraMineRInternalNodeInit}
\alias{TraMineRInternalSeqeage}
\alias{TraMineRInternalSeqgbar}
\alias{TraMineRInternalSplitInit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Access to TraMineR internal functions
}
\description{
Functions allowing other packages to access some TraMineR internal functions. Corresponding functions are respectively \code{TraMineR.setlayout}, \code{TraMineR.Legend}, \code{DTNInit}, \code{seqeage}, \code{seqgbar} and \code{DTNsplit}. For experts only.
}
\usage{
TraMineRInternalLayout(...)
TraMineRInternalLegend(...)
TraMineRInternalNodeInit(...)
TraMineRInternalSeqeage(...)
TraMineRInternalSeqgbar(...)
TraMineRInternalSplitInit(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{Arguments passed to or from other methods.}
}
